refactor(console): flatten composables/common into composables/
The three generic composables (useFrappeList, useFrappeRealtime, useIsMobile)
sat under common/ while ~30 domain composables lived flat beside them; the extra
directory earned nothing. Move them up and drop common/.
