+++ dashboard/src/composables/useIsMobile.ts
